Cleaned up ydkgen code
 * Do not catch YdkGenException when generating
