+++ ydkgen/printer/__init__.py
@@ -1,0 +1,38 @@
+#  ----------------------------------------------------------------
+# Copyright 2016 Cisco Systems
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ------------------------------------------------------------------
+
+"""
+printer.py
+ 
+ YANG model driven API, emitter.
+ 
+"""
+
+
+from .python.bits_printer import BitsPrinter
+from .python.class_common_path_printer import ClassCommonPathPrinter
+from .python.class_docstring_printer import ClassDocstringPrinter
+from .python.class_has_data_printer import ClassHasDataPrinter
+from .python.class_inits_printer import ClassInitsPrinter
+from .python.class_is_config_printer import ClassIsConfigPrinter
+from .python.class_meta_printer import ClassMetaPrinter
+from .python.class_printer import ClassPrinter
+from .python.enum_printer import EnumPrinter
+from .python.import_test_printer import ImportTestPrinter
+from .python.python_module_printer import PythonModulePrinter
+from .python.python_rst_printer import PythonRstPrinter
+from .python.test_case_printer import TestCasePrinter
+from .ydk_printer import YdkPrinter
