+++ ydkgen/printer/python/class_has_data_printer.py
