+++ ydkgen/printer/ydk_printer.py
@@ -20,18 +20,15 @@
 '''
 
 
-import api_model
-import common
-import gen_target
-import optparse
 import os
 import shutil
 
-from pyang import error
-from pyang import plugin
-from module_printer import PythonModulePrinter
-from helper import get_rst_file_name
-from api_model import Class, Enum
+from ydkgen import common
+from ydkgen.common import yang_id
+from ydkgen.common import convert_to_reStructuredText
+from .printer_factory import PrinterFactory, PrinterType
+from ydkgen.common import get_rst_file_name
+from ydkgen.api_model import Bits, Class, Enum
 
 
 class _EmitArgs:
@@ -45,11 +42,15 @@ class _EmitArgs:
 # ============================================================
 # Plugin class definitions
 # ============================================================
-class YDKPythonGen():
+class YdkPrinter():
 
-    def __init__(self, ydk_dir, ydk_doc_dir):
+    def __init__(self, ydk_dir, ydk_doc_dir, language):
         self.ydk_dir = ydk_dir
         self.ydk_doc_dir = ydk_doc_dir
+        if language == 'python':
+            self.language = PrinterType.PYTHON
+        else:
+            raise Exception('Language not supported')
 
     def emit(self, packages):
         self.ypy_ctx = None
@@ -81,7 +82,7 @@ class YDKPythonGen():
         self.ypy_ctx = common.PrintCtx()
         self.ypy_ctx.meta = True
         self.ypy_ctx.tab_size = 4
-        self.ypy_ctx.printer = PythonModulePrinter(self.ypy_ctx)
+        self.ypy_ctx.printer = PrinterFactory().get_printer(self.language)(self.ypy_ctx)
 
     def print_files(self):
         self.print_modules()
@@ -129,7 +130,7 @@ class YDKPythonGen():
 
         def _walk_n_print(named_element, p):
             self.print_file(get_python_rst_file_name(p, named_element),
-                            gen_target.emit_python_rst,
+                            emit_python_rst,
                             _EmitArgs(self.ypy_ctx, named_element))
 
             for owned_element in named_element.owned_elements:
@@ -144,7 +145,7 @@ class YDKPythonGen():
         packages = [p for p in self.packages if len(p.owned_elements) > 0]
 
         self.print_file(get_python_rst_ydk_models_file_name(self.ydk_doc_dir),
-                        gen_target.emit_ydk_models_rst,
+                        emit_ydk_models_rst,
                         _EmitArgs(self.ypy_ctx, packages))
 
     def print_python_module(self, package, index, path, size, sub):
@@ -152,36 +153,36 @@ class YDKPythonGen():
 
         package.parent_pkg_name = sub
         self.print_file(get_python_module_file_name(path, package),
-                        gen_target.emit_module,
+                        emit_module,
                         _EmitArgs(self.ypy_ctx, package, (sub, package.name)))
 
     def print_meta_module(self, package, path):
         self.print_init_file(path)
         self.print_file(get_meta_module_file_name(path, package),
-                        gen_target.emit_meta,
+                        emit_meta,
                         _EmitArgs(self.ypy_ctx, package))
 
     def print_test_module(self, package, path):
         self.print_file(get_test_module_file_name(path, package),
-                        gen_target.emit_test_module,
+                        emit_test_module,
                         _EmitArgs(self.ypy_ctx, package))
 
     def print_yang_ns_file(self):
         packages = self.packages + self.deviation_packages
         self.print_file(get_yang_ns_file_name(self.models_dir),
-                        gen_target.emit_yang_ns,
+                        emit_yang_ns,
                         _EmitArgs(self.ypy_ctx, packages))
 
     def print_deviate_file(self):
         self.print_init_file(self.deviation_dir)
         for package in self.deviation_packages:
             self.print_file(get_meta_module_file_name(self.deviation_dir, package),
-                            gen_target.emit_deviation,
+                            emit_deviation,
                             _EmitArgs(self.ypy_ctx, package))
 
     def print_import_tests_file(self):
         self.print_file(get_import_test_file_name(self.test_dir),
-                        gen_target.emit_importests,
+                        emit_importests,
                         _EmitArgs(self.ypy_ctx, self.packages))
 
     def print_init_file(self, path):
@@ -238,3 +239,154 @@ def get_meta_module_file_name(path, package):
 
 def get_test_module_file_name(path, package):
     return '%s/%sTest.py' % (path, package.stmt.arg.replace('-', '_'))
+
+
+
+def emit_comment(ctx, comment):
+    ctx.printer.comment(lines=comment.split('\n'))
+
+
+def emit_module_header(ctx, package, mheader=None, is_meta=False):
+    # ::::::::::::::::::::::::::::::::::::::::
+    # Print the header
+    # ::::::::::::::::::::::::::::::::::::::::
+    s = package.stmt
+    if is_meta:
+        rpcs = [idx for idx in package.owned_elements if isinstance(idx, Class) and idx.is_rpc()]
+        anyxml_import = ''
+        if len(rpcs) > 0:
+            anyxml_import = ', ANYXML_CLASS'
+        ctx.printer.meta_header(anyxml_import)
+    else:
+        comment = s.search_one('description')
+        ctx.writeln('""" %s ' % package.name)
+        ctx.bline()
+
+        if comment is not None and not is_meta:
+            ctx.comment = comment.arg
+            for line in ctx.comment.split('\n'):
+                ctx.writeln(convert_to_reStructuredText(line))
+        ctx.bline()
+        ctx.writeln('"""')
+        ctx.printer.header(mheader)
+        ctx.printer.imports(package)
+    ctx.root = s
+    ctx.augment_path = ''
+    ctx.aug_stmt = None
+    ctx.module_name = yang_id(s)
+    ctx.module = s
+    # get the yang meta information.
+    prefix = s.search_one('prefix')
+    if prefix is not None:
+        ctx.prefix = yang_id(prefix)
+    namespace = s.search_one('namespace')
+    if namespace is not None:
+        ctx.namespace = yang_id(namespace)
+    org = s.search_one('organization')
+    if org is not None:
+        ctx.organization = yang_id(org)
+    contact = s.search_one('contact')
+    if contact is not None:
+        ctx.contact = yang_id(contact)
+    revision = s.search_one('revision')
+    if revision is not None:
+        ctx.revision = yang_id(revision)
+
+    ctx.ns += [(yang_id(s), ctx.namespace)]
+
+
+def emit_yang_ns(ctx, packages):
+
+    ctx.printer.print_yang_ns_header()
+    ns_list = []
+    namespace_map = {}
+    for m in [p.stmt for p in packages]:
+
+        ns = m.search_one('namespace')
+        if ns is not None:
+            ns_list.append((m.arg.replace('-', '_'), ns.arg, yang_id(m)))
+
+    for package in packages:
+        ns = package.stmt.search_one('namespace')
+        for ele in package.owned_elements:
+            if hasattr(ele, 'stmt') and ele.stmt is not None and ele.stmt.keyword == 'container':
+                namespace_map[(ns.arg, ele.stmt.arg)] = (package.get_py_mod_name(), ele.name)
+
+
+    ctx.printer.print_namespaces(ns_list)
+    ctx.printer.print_identity_map(packages)
+    ctx.printer.print_namespaces_map(namespace_map)
+
+
+def emit_importests(ctx, packages):
+    ctx.printer.print_import_tests(packages)
+
+
+def emit_python_rst(ctx, named_element):
+    ctx.printer.print_python_rst(named_element)
+
+
+def emit_ydk_models_rst(ctx, packages):
+    ctx.printer.print_ydk_models_rst(packages)
+
+
+def emit_module(ctx, package, mheader):
+
+    emit_module_header(ctx, package, mheader=mheader)
+
+    if package is not None:
+        emit_module_enums(ctx, package)
+        emit_module_bits(ctx, package)
+        emit_module_classes(ctx, package)
+        ctx.bline()
+
+
+def emit_module_enums(ctx, package):
+    enumz = []
+    enumz.extend(
+        [element for element in package.owned_elements if isinstance(element, Enum)])
+    for nested_enumz in sorted(enumz, key=lambda element: element.name):
+        ctx.printer.print_enum(nested_enumz)
+
+
+def emit_module_bits(ctx, package):
+    bits = []
+    bits.extend(
+        [bit for bit in package.owned_elements if isinstance(bit, Bits)])
+    for bit in sorted(bits, key=lambda bit: bit.name):
+        ctx.printer.print_bits(bit)
+
+
+def emit_module_classes(ctx, package):
+    ctx.printer.print_classes_at_same_level(
+        [clazz for clazz in package.owned_elements if isinstance(clazz, Class)])
+
+
+def emit_test_module(ctx, package):
+    ctx.printer.print_testcases(package)
+
+
+def emit_meta(ctx, package):
+    ctx.print_meta = True
+    emit_module_header(ctx, package, is_meta=True)
+    if package is not None:
+        emit_meta_table_open(ctx)
+        for nested_enumz in [e for e in package.owned_elements if isinstance(e, Enum)]:
+            ctx.printer.print_enum_meta(nested_enumz)
+        ctx.printer.print_classes_meta([c for c in package.owned_elements if isinstance(c, Class)])
+        emit_meta_table_close(ctx)
+        ctx.printer.print_classes_meta_parents(
+            [c for c in package.owned_elements if isinstance(c, Class)])
+
+
+def emit_meta_table_open(ctx):
+    ctx.writeln('_meta_table = {')
+    ctx.lvl_inc()
+
+
+def emit_meta_table_close(ctx):
+    ctx.lvl_dec()
+    ctx.writeln('}')
+
+def emit_deviation(ctx, package):
+    ctx.printer.print_deviation(package)
